[#9616] fix(authz): Fix the  user verfication logic
