+++ server-common/src/main/java/org/apache/gravitino/server/authorization/PassThroughAuthorizer.java
@@ -20,11 +20,15 @@ package org.apache.gravitino.server.authorization;
 import java.io.IOException;
 import java.security.Principal;
 import org.apache.gravitino.Entity;
+import org.apache.gravitino.GravitinoEnv;
 import org.apache.gravitino.MetadataObject;
 import org.apache.gravitino.NameIdentifier;
+import org.apache.gravitino.authorization.AccessControlDispatcher;
 import org.apache.gravitino.authorization.AuthorizationRequestContext;
 import org.apache.gravitino.authorization.GravitinoAuthorizer;
 import org.apache.gravitino.authorization.Privilege;
+import org.apache.gravitino.exceptions.NoSuchUserException;
+import org.apache.gravitino.utils.PrincipalUtils;
 
 /**
  * The default implementation of GravitinoAuthorizer, indicating that metadata permission control is
@@ -76,6 +80,14 @@ public class PassThroughAuthorizer implements GravitinoAuthorizer {
 
   @Override
   public boolean isMetalakeUser(String metalake) {
+    AccessControlDispatcher dispatcher = GravitinoEnv.getInstance().accessControlDispatcher();
+    if (dispatcher != null) {
+      try {
+        dispatcher.getUser(metalake, PrincipalUtils.getCurrentUserName());
+      } catch (NoSuchUserException e) {
+        dispatcher.addUser(metalake, PrincipalUtils.getCurrentUserName());
+      }
+    }
     return true;
   }
